requirements.txt
torch>=1.11.0
transformers>=4.27.0
datasets>=2.10.0
pyyaml>=6.0
pandas>=1.5.0
matplotlib>=3.5.0
tqdm>=4.65.0
scikit-learn>=1.0.0
numpy>=1.22.0
google-cloud-storage>=2.7.0
google-cloud-aiplatform>=1.25.0 